add gkr_self_checks feature, refactor sumchek_loop, small fixes
